register controlller bugs

+++ src/utils/cloudinary.js
@@ -5,9 +5,9 @@ import fs from "fs"
 
     // Configuration
     cloudinary.config({ 
-        cloud_name: 'process.env.CLOUDINARY_CLOUD_NAME', 
-        api_key: 'process.env.CLOUDINARY_API_KEY', 
-        api_secret: 'process.env.CLOUDINARY_API_SECRET' // Click 'View API Keys' above to copy your API secret
+        cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
+        api_key: process.env.CLOUDINARY_API_KEY, 
+        api_secret: process.env.CLOUDINARY_API_SECRET  // Click 'View API Keys' above to copy your API secret
     });
     
 
@@ -21,15 +21,23 @@ import fs from "fs"
         
     
     //file uploaded successfully
-    console.log("file is uploaded",
-    response.url);
+    //console.log("file is uploaded",
+    //response.url);
+
+    fs.unlinkSync(localFilePath)
     return response;
      }
     catch(error){
-        fs.unlinkSync(localFilePath) 
+        console.error("cloudianry upload failed",error)
+        //fs.unlinkSync(localFilePath) 
+       if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        }
+
         return null;
             
     }
 }
+
 
 export {uploadOnCloudinary}
